requirements.txt
fastapi==0.95.0
loguru==0.6.0
peewee==3.16.0
pydantic==1.10.7
PyJWT==2.6.0
PyMySQL==1.0.3
starlette==0.26.1
toml==0.10.2
uvicorn==0.21.1
